[refactor] remove pi agg circuit; integrated into proof agg circuit
