+++ aggregator/src/tests.rs
@@ -2,7 +2,6 @@ pub(crate) mod aggregator_e2e;
 pub(crate) mod mock_chunk;
 pub(crate) mod proof_aggregation;
 pub(crate) mod proof_compression;
-pub(crate) mod public_input_aggregation;
 
 #[macro_export]
 macro_rules! layer_0 {
